Authorizing the user by creating jwt token

+++ controllers/api/v1/post_api.js
@@ -3,6 +3,7 @@ const Comment=require('../../../models/comment')
 
 module.exports.index=async function(req,res){
 
+    
     let posts = await Post.find({})
     .sort('-createdAt')      // THE post which was created earlier appear at the top and sorting occurs
     .populate('user')
@@ -21,21 +22,30 @@ module.exports.index=async function(req,res){
 }
 
 module.exports.destroy = async function (req, res) {
+    console.log("*****************************************")
 
     try {
-        let post = await Post.findById(req.params.id)
-
+        let posts = await Post.findById(req.params.id)
        
-            post.remove()
+        if(posts.user==req.user.id){
+
+            posts.remove()
             await Comment.deleteMany({ post: req.params.id })
 
            return res.json(200,{
             message:"post and comment associated with it deleted"
            })
+            
+        }else{
+            return res.json(401,{
+                message:"You Are Not Authorized To Delete the post"
+            })
+        }
+        
 
 
     } catch (err) {
-        console.log('Error', err)
+        console.log('Error*****************************************', err)
         return res.json(500,{
             message:"Internal Server Error"
         })
